Update routes and controllers for products
