+++ src/routes/products.js
@@ -3,7 +3,14 @@ const router = express.Router()
 
 const productController = require('../controller/products')
 
-router.get('/productCart', productController.renderProductCart)
-router.get('/productDetail/:id', productController.renderProductDetail)
+router.get('/cart', productController.renderProductCart)
+
+router.get('/create', productController.renderProductCreat)
+router.get('/:id', productController.renderProductDetail)
+
+router.post('/', productController.addProduct)
+router.get('/:id/edit', productController.renderProductEdit)
+
+router.put('/:id', productController.editProduct)
 
 module.exports = router
