+++ src/routes/user.js
@@ -5,10 +5,5 @@ const userController = require('../controller/user')
 
 router.get('/login', userController.renderLogin)
 router.get('/register', userController.renderRegister)
-router.get('/productCreat', userController.renderProductCreat)
-router.get('/productEdit/:id', userController.renderProductEdit)
-
-router.post('/productCreat', userController.adminAddProduct)
-router.post('/productEdit/:id', userController.adminEditProduct)
 
 module.exports = router
